requirements.txt
# Blocky Market Maker Bot - Dependencies

# Environment and configuration
python-dotenv>=1.0.0

# YAML configuration
PyYAML>=6.0

# HTTP requests (for alerts and API)
requests>=2.31.0

# WebSocket support
websockets>=12.0

# Async HTTP client
aiohttp>=3.9.0

# Configuration validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Prometheus metrics
prometheus-client>=0.19.0

# Dashboard templating
aiohttp-jinja2>=1.5.0
jinja2>=3.1.0

# System Tray (Windows/macOS/Linux)
pystray>=0.19.0
Pillow>=10.0.0

# Development (optional - for building Windows executable)
# pyinstaller>=6.0.0
